+++ src/page/handover/note/Custom/Custom.tsx
@@ -1,7 +1,7 @@
-import { ChangeEvent, useCallback, useRef, useState } from 'react';
+import { useCallback } from 'react';
 
 import Button from '@/common/component/Button/Button';
-import Text from '@/common/component/Text/Text';
+import Label from '@/common/component/Label/Label';
 import { scrollStyle } from '@/common/style/scroll';
 
 import {
@@ -13,24 +13,22 @@ import {
 } from '@/page/handover/note/Custom/Custom.style';
 import File from '@/page/handover/note/component/file/File';
 
-const Custom = () => {
-  const fileRef = useRef<HTMLInputElement>(null);
-  const [files, setFiles] = useState<File[]>([]);
+import useFile from '../hooks/useFile';
+
+interface CustomProps {
+  onSubmit: () => void;
+}
+
+const Custom = ({ onSubmit }: CustomProps) => {
+  const { files, handleFileChange } = useFile();
 
   const handleFileUpload = useCallback(() => {
-    fileRef.current?.click();
-  }, []);
-
-  const handleFileChange = useCallback((e: ChangeEvent<HTMLInputElement>) => {
-    const selectedFile = e.target.files?.[0];
-
-    if (selectedFile) {
-      setFiles((prev) => [...prev, selectedFile]);
-    }
+    const fileInput = document.getElementById('file') as HTMLInputElement;
+    fileInput?.click();
   }, []);
 
   return (
-    <div css={[noteWrapperStyle, scrollStyle]}>
+    <form css={[noteWrapperStyle, scrollStyle]} onSubmit={onSubmit}>
       <div css={layoutStyle}>
         <textarea
           css={textareaStyle}
@@ -38,10 +36,10 @@ const Custom = () => {
         />
       </div>
       <div css={layoutStyle}>
-        <Text tag="body4" css={guideStyle}>
+        <Label id="file" css={guideStyle}>
           드라이브에서 연동하고 싶은 파일을 선택해주세요.
-        </Text>
-        <input type="file" style={{ display: 'none' }} ref={fileRef} onChange={handleFileChange} />
+        </Label>
+        <input id="file" type="file" style={{ display: 'none' }} onChange={handleFileChange} />
         <div css={fileBoxStyle}>
           {files.map((file) => (
             <File key={file.name} file={file} />
@@ -51,7 +49,7 @@ const Custom = () => {
           파일 연동하기
         </Button>
       </div>
-    </div>
+    </form>
   );
 };
